refactor: replace test files in molecules
