+++ src/molecules/inputs/numeric/NumericInput.helpers.test.ts
@@ -1,7 +1,7 @@
 import { ChopLogicNumericInputValidator } from '@models';
 import { describe, expect, it } from 'vitest';
 
-import { getNumericInputInitialValue, validateNumericInputValue } from '../NumericInput.helpers.ts';
+import { getNumericInputInitialValue, validateNumericInputValue } from './NumericInput.helpers.ts';
 
 describe('getNumericInputInitialValue', () => {
   it('should return the value from initialValues if it exists', () => {
