+++ src/molecules/inputs/numeric/NumericInput.test.tsx
@@ -3,7 +3,7 @@ import { fireEvent, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, expect, it, vi } from 'vitest';
 
-import NumericInput from '../NumericInput.tsx';
+import NumericInput from './NumericInput.tsx';
 
 describe('NumericInput', () => {
   const testProps = {
