+++ src/molecules/inputs/text/TextInput.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, expect, it, vi } from 'vitest';
 
-import TextInput from '../TextInput.tsx';
+import TextInput from './TextInput.tsx';
 
 describe('ChopLogicTextInput', () => {
   const testProps = {
