+++ src/organisms/grid/Grid.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, expect, it, vi } from 'vitest';
 
-import ChopLogicGrid from '../Grid.tsx';
+import ChopLogicGrid from './Grid';
 
 describe('ChopLogicGrid', () => {
   const testProps = {
